// src/app/model/createVendor.ts
export class createVendor {
    vendorName: string | undefined;
    vendorLicenseOwner: string | undefined;
    registrationNumber: string | undefined;
    gstNumber: string | undefined;
    vendorAddress: string | undefined;
    state: string | undefined;
    city: string | undefined;
    postalCode: string | undefined;
    phoneNumber: string | undefined;
    vendorUsername: string | undefined;
    role: string | undefined;
}